requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
mysql-connector-python==8.2.0
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
